Bug 4517 error: comparison between signed and unsigned integer
The old cast is required when size_t is unsigned (as it should be).
The new cast is required when size_t is signed (as it may be).

We could cast just the left-hand side to be signed instead, but it feels
slightly wrong to do that here because all values we are casting are
meant to be unsigned and, hence, in theory, might overflow in some
future version of the code if we cast them to a signed value now and
forget to fix that cast later while adding support for larger values.

+++ src/fs/rock/RockIoState.cc
@@ -185,7 +185,7 @@ Rock::IoState::tryWrite(char const *buf, size_t size, off_t coreOff)
     assert(!coreOff || coreOff == -1);
 
     // throw if an accepted unknown-size entry grew too big or max-size changed
-    Must(offset_ + size <= static_cast<uint64_t>(dir->maxObjectSize()));
+    Must(static_cast<uint64_t>(offset_ + size) <= static_cast<uint64_t>(dir->maxObjectSize()));
 
     // allocate the first slice during the first write
     if (!coreOff) {
